jz4740: move JZ4740 SoC support into a subdirectory
This is in preparation for the addition of support for further XBurst
SoCs which will exist in other subdirectories of the xburst directory.

SoC-specific include files are moved under an arch-${SOC} directory and
accessed via the asm/arch symlink, and an asm/arch/jzsoc.h include file
is added to include jz4740.h without referencing the jz4740
specifically. This will allow further XBurst SoCs to have their jzsoc.h
header include their own SoC header(s). Drivers or other code specific
to the XBurst family but not to a single XBurst SoC can then include
jzsoc.h to access register definitions, addresses etc without having to
know or care about which SoC is in use.