ns16550: support for Ingenic JZ47xx SoCs
Ingenic JZ47xx series SoCs include a UART block which is almost NS16550
compatible. However it requires that a bit be set in the FCR register in
order to enable the UART block. Setting this bit will allow the generic
NS16550 driver to function on Ingenic JZ47xx series SoCs and allow the
removal of an Ingenic-specific serial driver.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>