HACK: nand: check explicitly for 0x00 as the factory BBM rather than != 0xff
This is a hack to avoid any potential problems with factory bad block
marker identification. The NAND chip that we have uses a zero byte in
the first OOB location of the first page of a bad block as the BBM (this
is also in line with Factory Defect Mapping requirements in the the ONFI
specification). However U-Boot will take any byte that is not 0xff to be
a bad block.

While the chips should be factory erased, Ingenic have been testing on
chips that are not erased, and U-Boot mistakenly identifies most of the
blocks on the chip as bad. Avoid this by testing explicitly for a zero
byte.